Factoriser les métriques partagées et uniformiser les cartes de soldes
Agent-Logs-Url: https://github.com/Cyril-CB/CS-PILOT/sessions/b85bea9a-a491-4134-a9a4-064a372318e3
