+++ blueprints/suivi.py
@@ -5,7 +5,14 @@ from flask import Blueprint, render_template, request, redirect, url_for, sessio
 from datetime import datetime, timedelta
 import json
 from database import get_db
-from utils import login_required, get_semaine_alternance
+from utils import login_required
+from blueprints.worktime_metrics import (
+    BUSINESS_DAYS_PER_WEEK,
+    compute_day_metrics as _compute_day_metrics,
+    get_feries_set as _get_feries_set,
+    get_last_completed_week_monday as _get_last_completed_week_monday,
+    recent_business_days as _recent_business_days,
+)
 
 suivi_bp = Blueprint('suivi_bp', __name__)
 
@@ -17,11 +24,8 @@ SURCHARGE_CATEGORIES = [
     {'label': 'Rouge', 'min': 76, 'max': 99, 'color': '#dc2626', 'text_color': '#7f1d1d'},
     {'label': 'Noir', 'min': 100, 'max': 100, 'color': '#111827', 'text_color': '#111827'},
 ]
-MIN_BREAK_MINUTES = 20
 MIN_DAILY_REST_HOURS = 11
 MAX_CONSECUTIVE_HOURS_WITH_SHORT_BREAK = 6
-BUSINESS_DAYS_PER_WEEK = 5
-SATURDAY_WEEKDAY_INDEX = 5
 MAX_DAILY_HOURS_THRESHOLD = 10
 MAX_WEEKLY_AVERAGE_HOURS = 44
 MAX_WEEKLY_HOURS_THRESHOLD = 48
@@ -52,181 +56,6 @@ def _get_previous_month_range(reference_date):
     last_day_previous_month = first_day_current_month - timedelta(days=1)
     first_day_previous_month = last_day_previous_month.replace(day=1)
     return first_day_previous_month, last_day_previous_month
-
-
-def _get_last_completed_week_monday(reference_date):
-    current_week_monday = reference_date - timedelta(days=reference_date.weekday())
-    return current_week_monday - timedelta(days=7)
-
-
-def _get_feries_set(conn):
-    return {
-        row['date']
-        for row in conn.execute('SELECT date FROM jours_feries').fetchall()
-    }
-
-
-def _get_type_periode_cached(conn, date_str, type_periode_cache):
-    if date_str not in type_periode_cache:
-        periode = conn.execute('''
-            SELECT 1
-            FROM periodes_vacances
-            WHERE ? >= date_debut AND ? <= date_fin
-            LIMIT 1
-        ''', (date_str, date_str)).fetchone()
-        type_periode_cache[date_str] = 'vacances' if periode else 'periode_scolaire'
-    return type_periode_cache[date_str]
-
-
-def _get_planning_cached(conn, user_id, date_str, planning_cache, type_periode_cache):
-    cache_key = (user_id, date_str)
-    if cache_key not in planning_cache:
-        type_periode = _get_type_periode_cached(conn, date_str, type_periode_cache)
-        semaine_type = get_semaine_alternance(user_id, date_str)
-        if semaine_type == 'fixe':
-            planning = conn.execute('''
-                SELECT *
-                FROM planning_theorique
-                WHERE user_id = ?
-                  AND type_periode = ?
-                  AND (type_alternance IS NULL OR type_alternance = 'fixe')
-                  AND date_debut_validite <= ?
-                ORDER BY date_debut_validite DESC
-                LIMIT 1
-            ''', (user_id, type_periode, date_str)).fetchone()
-        else:
-            planning = conn.execute('''
-                SELECT *
-                FROM planning_theorique
-                WHERE user_id = ?
-                  AND type_periode = ?
-                  AND type_alternance = ?
-                  AND date_debut_validite <= ?
-                ORDER BY date_debut_validite DESC
-                LIMIT 1
-            ''', (user_id, type_periode, semaine_type, date_str)).fetchone()
-        planning_cache[cache_key] = planning
-    return planning_cache[cache_key]
-
-
-def _day_segments_from_row(row):
-    segments = []
-    if row['heure_debut_matin'] and row['heure_fin_matin']:
-        segments.append((row['heure_debut_matin'], row['heure_fin_matin']))
-    if row['heure_debut_aprem'] and row['heure_fin_aprem']:
-        segments.append((row['heure_debut_aprem'], row['heure_fin_aprem']))
-    return segments
-
-
-def _day_segments_from_planning(planning, date_obj):
-    if not planning or date_obj.weekday() >= BUSINESS_DAYS_PER_WEEK:
-        return []
-
-    jour_nom = ['lundi', 'mardi', 'mercredi', 'jeudi', 'vendredi'][date_obj.weekday()]
-    segments = []
-    matin_debut = planning[f'{jour_nom}_matin_debut']
-    matin_fin = planning[f'{jour_nom}_matin_fin']
-    aprem_debut = planning[f'{jour_nom}_aprem_debut']
-    aprem_fin = planning[f'{jour_nom}_aprem_fin']
-
-    if matin_debut and matin_fin:
-        segments.append((matin_debut, matin_fin))
-    if aprem_debut and aprem_fin:
-        segments.append((aprem_debut, aprem_fin))
-    return segments
-
-
-def _compute_segments_metrics(date_obj, segments):
-    metrics = {
-        'worked_hours': 0.0,
-        'start_at': None,
-        'end_at': None,
-        'break_minutes': None,
-        'longest_consecutive_hours': 0.0,
-    }
-    if not segments:
-        return metrics
-
-    dt_segments = []
-    for start_str, end_str in segments:
-        start_dt = datetime.combine(date_obj, datetime.strptime(start_str, '%H:%M').time())
-        end_dt = datetime.combine(date_obj, datetime.strptime(end_str, '%H:%M').time())
-        if end_dt < start_dt:
-            end_dt += timedelta(days=1)
-        dt_segments.append((start_dt, end_dt))
-
-    metrics['worked_hours'] = round(sum((end - start).total_seconds() for start, end in dt_segments) / 3600, 2)
-    metrics['start_at'] = dt_segments[0][0]
-    metrics['end_at'] = dt_segments[-1][1]
-
-    if len(dt_segments) >= 2:
-        break_seconds = (dt_segments[1][0] - dt_segments[0][1]).total_seconds()
-        metrics['break_minutes'] = round(max(0, break_seconds) / 60, 2)
-
-    longest_seconds = (dt_segments[0][1] - dt_segments[0][0]).total_seconds()
-    current_seconds = longest_seconds
-    for index in range(1, len(dt_segments)):
-        gap_seconds = (dt_segments[index][0] - dt_segments[index - 1][1]).total_seconds()
-        segment_seconds = (dt_segments[index][1] - dt_segments[index][0]).total_seconds()
-        if gap_seconds < MIN_BREAK_MINUTES * 60:
-            current_seconds += max(gap_seconds, 0) + segment_seconds
-        else:
-            longest_seconds = max(longest_seconds, current_seconds)
-            current_seconds = segment_seconds
-    longest_seconds = max(longest_seconds, current_seconds)
-    metrics['longest_consecutive_hours'] = round(longest_seconds / 3600, 2)
-
-    return metrics
-
-
-def _compute_day_metrics(conn, user_id, row, planning_cache, type_periode_cache):
-    date_obj = datetime.strptime(row['date'], '%Y-%m-%d').date()
-    planning = _get_planning_cached(conn, user_id, row['date'], planning_cache, type_periode_cache)
-    planned_segments = _day_segments_from_planning(planning, date_obj)
-    actual_segments = planned_segments if row['declaration_conforme'] else _day_segments_from_row(row)
-
-    planned_metrics = _compute_segments_metrics(date_obj, planned_segments)
-    actual_metrics = _compute_segments_metrics(date_obj, actual_segments)
-
-    theoretical_hours = 0.0
-    if date_obj.weekday() == SATURDAY_WEEKDAY_INDEX:
-        theoretical_hours = 0.0
-    elif date_obj.weekday() < BUSINESS_DAYS_PER_WEEK:
-        theoretical_hours = planned_metrics['worked_hours']
-
-    actual_hours = theoretical_hours if row['declaration_conforme'] else actual_metrics['worked_hours']
-
-    pause_reduced = (
-        planned_metrics['break_minutes'] is not None
-        and actual_metrics['break_minutes'] is not None
-        and actual_metrics['break_minutes'] < planned_metrics['break_minutes']
-    )
-
-    return {
-        'date': row['date'],
-        'date_obj': date_obj,
-        'theoretical_hours': theoretical_hours,
-        'actual_hours': actual_hours,
-        'delta': round(actual_hours - theoretical_hours, 2),
-        'start_at': actual_metrics['start_at'],
-        'end_at': actual_metrics['end_at'],
-        'break_minutes': actual_metrics['break_minutes'],
-        'planned_break_minutes': planned_metrics['break_minutes'],
-        'pause_reduced': pause_reduced,
-        'longest_consecutive_hours': actual_metrics['longest_consecutive_hours'],
-    }
-
-
-def _recent_business_days(end_date, limit, feries_set):
-    days = []
-    current = end_date
-    while len(days) < limit:
-        current_str = current.strftime('%Y-%m-%d')
-        if current.weekday() < BUSINESS_DAYS_PER_WEEK and current_str not in feries_set:
-            days.append(current_str)
-        current -= timedelta(days=1)
-    days.reverse()
-    return days
 
 
 def _format_points(label, points, detail):
